Reduce indirection in unpacking a response
Now instance variables are set directly from the packet and the mode is
unpacked directly, and from an integer instead of a string.

This trades some overhead in directly setting values from the first byte
for less overhead when reading values outside the first byte.

This also will make it easier to pack a packet back into a String as
everything is more explicit.